Add -MOVED and -ASK redirects
From now on, only the instance owning a key (or slot) will reply to GET
and SET requests. The other instances will reply with a redirect.

ASK redirects are implemented as exceptions to the slot mapping. They
are created using fakeredis_cluster:set_ask_redirect/3.

Some function are added to check the slot mapping.